pyproject.toml
[tool.poetry]
name = "wwi_thor_data_api"
version = "0.1.0"
description = ""
authors = ["Dario Thornhill <dariothornhill@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django-environ = "^0.4.5"
django-cors-headers = "3.7.0"
asgiref = "3.3.1"
Django = "3.1.7"
djangorestframework-simplejwt = "4.6.0"
djangorestframework = "3.12.2"
djangorestframework-gis = "0.17"
gunicorn = "20.0.4"
PyJWT = "2.0.1"
pytz = "2021.1"
sqlparse = "0.4.1"
whitenoise = "5.2.0"
pandas = "1.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
